Almost complete mapping xml->persistance->response

+++ src/main/java/no/regnskap/model/xml/RegnskapXmlInfo.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class RegnskapXmlInfo {
     private String feltkode;
-    private Long sum;
+    private BigDecimal sum;
 
     @JacksonXmlElementWrapper(useWrapping = false)
     private RegnskapXmlPost[] post;
@@ -18,7 +18,7 @@ public class RegnskapXmlInfo {
         return feltkode;
     }
 
-    public Long getSum() {
+    public BigDecimal getSum() {
         return sum;
     }
 
@@ -31,7 +31,7 @@ public class RegnskapXmlInfo {
     }
 
     public void setSum(BigDecimal sum) {
-        this.sum = sum.longValue();
+        this.sum = sum;
     }
 
     public void setPost(RegnskapXmlPost[] post) {
